requirements.txt
# Python dependencies for Tortoise TTS Voice Cloning Server
torch>=2.0.0
torchaudio>=2.0.0
tortoise-tts
flask>=2.0.0
flask-cors>=4.0.0
numpy>=1.21.0
scipy>=1.7.0
librosa>=0.9.0
soundfile>=0.10.0